Rename UserDefaultsClient to UserDefaultsDataSource
