+++ LokiPackage/Sources/Core/UserDefaults/UserDefaultsError.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+public enum UserDefaultsError: LocalizedError {
+    case missingValue(key: UserDefaultsKey)
+
+    public var errorDescription: String? {
+        switch self {
+        case .missingValue:
+            return L10n.valueDoesNotExistForTheKey
+        }
+    }
+}
